VW-927 Bug/Fix Change datatype CoordinateX&Y

+++ Infrastructure/EntityFramework/ManufacturingDB/ProductionStep.cs
@@ -28,8 +28,8 @@ namespace VErp.Infrastructure.EF.ManufacturingDB
         public DateTime? DeletedDatetimeUtc { get; set; }
         public int SortOrder { get; set; }
         public bool? IsGroup { get; set; }
-        public int? CoordinateX { get; set; }
-        public int? CoordinateY { get; set; }
+        public decimal? CoordinateX { get; set; }
+        public decimal? CoordinateY { get; set; }
         public int SubsidiaryId { get; set; }
 
         public virtual Step Step { get; set; }
